check deletion of document if owner delete or else fail

+++ HousingBoardAdministration.HousingAdministrationWeb/Pages/Meeting/Document/Delete.cshtml.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
@@ -8,15 +9,20 @@ namespace HousingBoardAdministration.HousingAdministrationWeb.Pages.Meeting.Docu
     public class DeleteModel : PageModel
     {
         private readonly IBffClient _bffClient;
+        private readonly UserManager<IdentityUser> _userManager;
 
-        public DeleteModel(IBffClient bffClient)
+
+        public DeleteModel(IBffClient bffClient, UserManager<IdentityUser> userManager)
         {
             this._bffClient = bffClient;
+            _userManager = userManager;
         }
         public static Guid MeetingId { get; set; }
 
         [BindProperty]
         public DocumentDeleteViewModel DocumentModel { get; set; } = new();
+
+        public static Guid DocumentOwnerId { get; set; } = new();
         public async Task<IActionResult> OnGet(Guid meetingId, Guid documentId)
         {
             MeetingId = meetingId;
@@ -26,6 +32,8 @@ namespace HousingBoardAdministration.HousingAdministrationWeb.Pages.Meeting.Docu
             {
                 Id = dto.Id
             };
+
+            DocumentOwnerId = dto.UserOwnerId;
             return Page();
         }
 
@@ -33,7 +41,15 @@ namespace HousingBoardAdministration.HousingAdministrationWeb.Pages.Meeting.Docu
         {
             if (!ModelState.IsValid) return Page();
 
-            await _bffClient.DeleteDocumentAsync(DocumentModel);
+            if(DocumentOwnerId == Guid.Parse(_userManager.GetUserId(User)))
+            {
+                await _bffClient.DeleteDocumentAsync(DocumentModel);
+            }
+            else
+            {
+                TempData["ErrorMes"] = "Du kan kun slette dine egne dokumenter!";
+                return Page();
+            }
 
             return RedirectToPage("/Meeting/Get", new { id = MeetingId });
         }
